Enter Tournament now correctly returns a Tournament DTO instead of A TeamDTO

+++ src/main/java/net/htlgkr/pos3/kainzt/SplatournamentServer/dtos/TournamentDTO.java
@@ -7,6 +7,6 @@ public record TournamentDTO(
     String name,
     String createdBy,
     TournamentStyle style,
-    int teamCount
+    Long teamCount
 ) {
 }
